requirements.txt
streamlit==1.35.0
google-generativeai==0.7.0
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
plotly==5.20.0
matplotlib==3.8.3
networkx==3.2.1
Pillow==10.2.0
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
pymongo==4.6.1
